add pipeline component sample with primitive output
