+++ sdk/python/jobs/pipelines/1b_pipeline_with_python_function_components/src/components.py
@@ -3,7 +3,8 @@ from random import randint
 from uuid import uuid4
 
 # mldesigner package contains the command_component which can be used to define component from a python function
-from mldesigner import command_component, Input, Output
+from mldesigner import command_component, Input, Output, Meta
+from typing_extensions import Annotated
 
 
 @command_component()
@@ -89,3 +90,19 @@ def eval_model(
     # Evaluate the incoming scoring result and output evaluation result.
     # Here only output a dummy file for demo.
     (Path(eval_output) / "eval_result").write_text("eval_result")
+
+
+# int, float, string, bool are used the same way when define a primary output.
+@command_component()
+def component_return_annotated_output(input_int: int) -> Annotated[int, Meta(description="test annotation int")]:
+    return 1
+
+
+@command_component()
+def component_return_int_output(input_int: int) -> int:
+    return input_int
+
+
+@command_component()
+def component_return_integer_output(input_int: int) -> Output(type="integer"):
+    return input_int
