--- sdk/python/jobs/pipelines/pipeline_with_components_use_primitive_output/command_component_with_primitive_output.yaml
@@ -1,37 +1,0 @@
-$schema: https://azuremlschemas.azureedge.net/latest/commandComponent.schema.json
-type: command
-
-name: optional_input
-version: 0.0.1
-display_name: test_optional_input_output
-description: A dummy scoring component
-
-inputs:
-  input_int:
-    type: integer
-    default: 1
-  input_float:
-    type: number
-    default: 1.1
-  input_str:
-    type: string
-    default: str
-  input_bool:
-    type: boolean
-    default: true
-outputs:
-  output_int:
-    type: integer
-  output_float:
-    type: number
-  output_str:
-    type: string
-  output_bool:
-    type: boolean
-code: ./
-command: >-
-  python primitive.py --input-int ${{inputs.input_int}} --input-float ${{inputs.input_float}}
-  --input-str ${{inputs.input_str}} --input-bool ${{inputs.input_bool}}
-  --output-int ${{outputs.output_int}} --output-float ${{outputs.output_float}}
-  --output-str ${{outputs.output_str}} --output-bool ${{outputs.output_bool}}
-environment: azureml://registries/azureml/environments/mldesigner-minimal/labels/latest
